auto_care_backend/requirements.txt
# Core Django
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# Async / Realtime
channels==4.3.1
channels-redis==4.3.0
daphne==4.0.0

# Task Queue
celery==5.5.3
redis>=4.6,<7

# Database
psycopg2-binary==2.9.10
dj-database-url==2.1.0

# Environment / Config
django-environ==0.11.2
python-dotenv==1.0.0

# CORS
django-cors-headers==4.3.1

# File handling
Pillow==10.2.0

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Development (optional)
ipython==8.20.0